requirements.txt
torch>=2.1
torchvision>=0.16
timm>=0.9
openslide-python>=1.3
Pillow>=10.0
numpy>=1.26
matplotlib>=3.8
scikit-image>=0.22
tqdm>=4.66
pandas>=2.1
scikit-learn>=1.3
scipy>=1.11
streamlit>=1.55
pymupdf>=1.26
